notifications cleanup
